+++ src/Kupo/App/Database/Types.hs
@@ -1,4 +1,5 @@
 {-# LANGUAGE CPP #-}
+{-# LANGUAGE DuplicateRecordFields #-}
 
 module Kupo.App.Database.Types 
   (
@@ -9,6 +10,10 @@ module Kupo.App.Database.Types
     -- * Setup
     , ConnectionType (..)
     , DBPool (..)
+
+    -- * Tracer
+    , TraceDatabase (..)
+    , TraceConnection (..)
     )
     where
 
@@ -45,6 +50,16 @@ import Kupo.Data.Pattern
 
 import Kupo.Prelude
 
+import Data.Severity
+    ( HasSeverityAnnotation
+    , Severity (..)
+    )
+import Kupo.Control.MonadLog
+    ( HasSeverityAnnotation (..)
+    )
+import Kupo.Control.MonadTime
+    ( DiffTime
+    )
 import qualified Kupo.Data.Database as DB
 
 data ConnectionType = ReadOnly | ReadWrite | WriteOnly
@@ -167,5 +182,156 @@ data DBPool m = DBPool {
   tryWithDatabase :: forall res. ConnectionType -> (Database m -> m res) -> m (Maybe res)
   , withDatabase :: forall res. ConnectionType -> (Database m -> m res) -> m res
   , withDatabaseExclusiveWriter :: forall a. DeferIndexesInstallation -> (Database m -> m a) -> m a
+  , maxConcurrentReaders :: Int
+  , maxConcurrentWriters :: Int
   , destroyResources :: m ()
 }
+
+
+--
+-- Tracer
+--
+
+data TraceDatabase where
+    DatabaseCreateNew
+        :: { filePath :: FilePath }
+        -> TraceDatabase
+    DatabaseConnection
+        :: { message :: TraceConnection }
+        -> TraceDatabase
+    DatabaseCurrentVersion
+        :: { currentVersion :: Int }
+        -> TraceDatabase
+    DatabaseNoMigrationNeeded
+        :: TraceDatabase
+    DatabaseRunningMigration
+        :: { from :: Int, to :: Int }
+        -> TraceDatabase
+    DatabaseCreateIndex
+        :: { newIndex :: Text }
+        -> TraceDatabase
+    DatabaseIndexAlreadyExists
+        :: { index :: Text }
+        -> TraceDatabase
+    DatabaseDeferIndexes
+        :: { warning :: Text }
+        -> TraceDatabase
+    DatabaseRunningInMemory
+        :: TraceDatabase
+    DatabasePathMustBeDirectory
+        :: { hint :: Text }
+        -> TraceDatabase
+    DatabaseMustBeLocal
+        :: { errorMessage :: Text }
+        -> TraceDatabase
+    DatabaseCloneSourceDatabase
+        :: TraceDatabase
+    DatabaseCleanupOldData
+        :: { table :: Text }
+        -> TraceDatabase
+    DatabaseImportTable
+        :: { table :: Text, pattern :: Text }
+        -> TraceDatabase
+    DatabaseImported
+        :: { rows :: Integer }
+        -> TraceDatabase
+    DatabaseRemoveIncompleteCopy
+        :: { filePath :: FilePath }
+        -> TraceDatabase
+    DatabaseCopyFinalize
+        :: TraceDatabase
+    DatabaseDebug
+        :: Text
+        -> TraceDatabase
+    deriving stock (Generic, Show)
+
+instance ToJSON TraceDatabase where
+    toEncoding =
+        defaultGenericToEncoding
+
+instance HasSeverityAnnotation TraceDatabase where
+    getSeverityAnnotation = \case
+        DatabaseCreateNew{}            -> Notice
+        DatabaseConnection { message } -> getSeverityAnnotation message
+        DatabaseCurrentVersion{}       -> Info
+        DatabaseNoMigrationNeeded{}    -> Debug
+        DatabaseRunningMigration{}     -> Notice
+        DatabaseRunningInMemory{}      -> Warning
+        DatabaseCreateIndex{}          -> Notice
+        DatabaseIndexAlreadyExists{}   -> Debug
+        DatabasePathMustBeDirectory{}  -> Error
+        DatabaseMustBeLocal{}          -> Error
+        DatabaseDeferIndexes{}         -> Warning
+        DatabaseCloneSourceDatabase{}  -> Notice
+        DatabaseCleanupOldData{}       -> Info
+        DatabaseImportTable{}          -> Notice
+        DatabaseImported{}             -> Info
+        DatabaseRemoveIncompleteCopy{} -> Notice
+        DatabaseCopyFinalize{}         -> Notice
+        DatabaseDebug{}                -> Warning
+
+data TraceConnection where
+    ConnectionCreateShortLived
+        :: { mode :: ConnectionType }
+        -> TraceConnection
+    ConnectionDestroyShortLived
+        :: { mode :: ConnectionType }
+        -> TraceConnection
+    ConnectionLocked
+        :: { attempts :: Word, retryingIn :: DiffTime }
+        -> TraceConnection
+    ConnectionBusy
+        :: { attempts :: Word, retryingIn :: DiffTime }
+        -> TraceConnection
+    ConnectionBeginQuery
+        :: { beginQuery :: LText }
+        -> TraceConnection
+    ConnectionExitQuery
+        :: { exitQuery :: LText }
+        -> TraceConnection
+    ConnectionCreateTemporaryIndex
+        :: { newTemporaryIndex :: Text }
+        -> TraceConnection
+    ConnectionCreatedTemporaryIndex
+        :: { newTemporaryIndex :: Text }
+        -> TraceConnection
+    ConnectionRemoveTemporaryIndex
+        :: { indexName :: Text }
+        -> TraceConnection
+    ConnectionRemoveIndex
+        :: { indexName :: Text }
+        -> TraceConnection
+    ConnectionFailedPragma
+        :: { pragma :: Text }
+        -> TraceConnection
+    ConnectionFailedToOpenDatabase
+        :: { hint :: Text }
+        -> TraceConnection
+    deriving stock (Generic, Show)
+
+instance ToJSON TraceConnection where
+    toEncoding =
+        defaultGenericToEncoding
+
+instance HasSeverityAnnotation TraceConnection where
+    getSeverityAnnotation = \case
+        ConnectionCreateShortLived{} -> Debug
+        ConnectionDestroyShortLived{} -> Debug
+        ConnectionLocked{attempts, retryingIn}
+            | retryingIn * fromIntegral attempts > 60  -> Warning
+        ConnectionLocked{} -> Debug
+        ConnectionBusy{attempts, retryingIn}
+            | retryingIn * fromIntegral attempts > 60 -> Warning
+        ConnectionBusy{} -> Debug
+        ConnectionBeginQuery{beginQuery}
+            | beginQuery == "PRAGMA optimize" -> Notice
+        ConnectionBeginQuery{} -> Debug
+        ConnectionExitQuery{exitQuery}
+            | exitQuery == "PRAGMA optimize" -> Notice
+        ConnectionExitQuery{} -> Debug
+        ConnectionCreateTemporaryIndex{} -> Debug
+        ConnectionCreatedTemporaryIndex{} -> Debug
+        ConnectionRemoveTemporaryIndex{} -> Debug
+        ConnectionRemoveIndex{} -> Warning
+        ConnectionFailedPragma{} -> Warning
+        ConnectionFailedToOpenDatabase{} -> Error
