+++ test/Test/Kupo/OptionsSpec.hs
@@ -29,8 +29,8 @@ import Kupo.Data.Cardano
 import Kupo.Data.Configuration
     ( ChainProducer (..)
     , Configuration (..)
+    , DatabaseLocation (..)
     , InputManagement (..)
-    , WorkDir (..)
     )
 import Kupo.Data.Pattern
     ( MatchBootstrap (..)
@@ -100,7 +100,7 @@ spec = parallel $ do
                 { nodeSocket = "./node.socket"
                 , nodeConfig = "./node.config"
                 }
-            , workDir = InMemory
+            , databaseLocation = InMemory
             }
           )
         , ( defaultArgs'
@@ -109,7 +109,7 @@ spec = parallel $ do
                 { ogmiosHost = "localhost"
                 , ogmiosPort = 1337
                 }
-            , workDir = InMemory
+            , databaseLocation = InMemory
             }
           )
         , ( defaultArgs''
@@ -118,14 +118,14 @@ spec = parallel $ do
                 { hydraHost = "localhost"
                 , hydraPort = 4001
                 }
-            , workDir = InMemory
+            , databaseLocation = InMemory
             }
           )
         , ( filter (/= "--in-memory") defaultArgs ++
             [ "--workdir", "./workdir"
             ]
           , shouldParseAppConfiguration $ defaultConfiguration
-            { workDir = Dir "./workdir"
+            { databaseLocation = Dir "./workdir"
             }
           )
         , ( defaultArgs ++ [ "--host", "0.0.0.0" ]
